Refactor Updater tests
Refactor Updater tests to be in one file, now that the
subclasses have been removed.
